move byteutils into util package
