+++ core/src/main/java/tech/pegasys/eth2signer/core/util/ByteUtils.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.eth2signer.core.utils;
+package tech.pegasys.eth2signer.core.util;
 
 import java.math.BigInteger;
